@@ -1,5 +1,4 @@
 import json
-from typing import Optional
 
 import aiohttp
 from aiohttp import web
@@ -16,7 +15,7 @@ from udata_hydra.utils import get_request_params
 async def get_latest_check(request: web.Request) -> web.Response:
     """Get the latest check for a given URL or resource_id"""
     url, resource_id = get_request_params(request, params_names=["url", "resource_id"])
-    data: Optional[dict] = await Check.get_latest(url, resource_id)
+    data: dict | None = await Check.get_latest(url, resource_id)
     if not data:
         raise web.HTTPNotFound()
     if data["deleted"]:
@@ -26,7 +25,7 @@ async def get_latest_check(request: web.Request) -> web.Response:
 
 async def get_all_checks(request: web.Request) -> web.Response:
     url, resource_id = get_request_params(request, params_names=["url", "resource_id"])
-    data: Optional[list] = await Check.get_all(url, resource_id)
+    data: list | None = await Check.get_all(url, resource_id)
     if not data:
         raise web.HTTPNotFound()
     return web.json_response([CheckSchema().dump(dict(r)) for r in data])
@@ -61,7 +60,7 @@ async def create_check(request: web.Request) -> web.Response:
         )
         context.monitor().refresh(status)
 
-    check: Optional[dict] = await Check.get_latest(url, resource_id)
+    check: dict | None = await Check.get_latest(url, resource_id)
     if not check:
         raise web.HTTPBadRequest(text=f"Check not created, status: {status}")
 
