@@ -1,8 +1,7 @@
 import json
-from typing import Tuple
 
 
-async def detect_tabular_from_headers(check: dict) -> Tuple[bool, str]:
+async def detect_tabular_from_headers(check: dict) -> tuple[bool, str]:
     """
     Determine from content-type header if file looks like:
         - a csv
