@@ -2,7 +2,7 @@ import gzip
 import hashlib
 import logging
 import tempfile
-from typing import IO, Optional
+from typing import IO
 
 import aiohttp
 import magic
@@ -33,7 +33,7 @@ def read_csv_gz(file_path: str) -> IO[bytes]:
 async def download_resource(
     url: str,
     headers: dict,
-    max_size_allowed: Optional[int],
+    max_size_allowed: int | None,
 ) -> IO[bytes]:
     """
     Attempts downloading a resource from a given url.
