@@ -1,9 +1,8 @@
 import importlib.metadata
 import logging
 import os
+import tomllib
 from pathlib import Path
-
-import toml
 
 log = logging.getLogger("udata-hydra")
 
@@ -19,12 +18,14 @@ class Configurator:
 
     def configure(self) -> None:
         # load default settings
-        configuration: dict = toml.load(Path(__file__).parent / "config_default.toml")
+        with open(Path(__file__).parent / "config_default.toml", "rb") as f:
+            configuration: dict = tomllib.load(f)
 
         # override with local settings
         local_settings = os.environ.get("HYDRA_SETTINGS", Path.cwd() / "config.toml")
         if Path(local_settings).exists():
-            configuration.update(toml.load(local_settings))
+            with open(Path(local_settings), "rb") as f:
+                configuration.update(tomllib.load(f))
 
         self.configuration = configuration
         self.check()
